requirements.txt
# requirements/common.txt
Django 				==	1.7.1
djangorestframework	==	2.3.13
Markdown			==	2.4.1
django-filter 		==	0.7
django-rest-swagger == 0.1.14
